+++ examples/use_locale/src/main.rs
@@ -1,13 +1,14 @@
 use leptos::*;
 use leptos_use::docs::demo_or_body;
 use leptos_use::use_locale;
+use unic_langid::langid_slice;
 
 #[component]
 fn Demo() -> impl IntoView {
-    let locale = use_locale(["en", "de", "fr"]);
+    let locale = use_locale(langid_slice!["en", "de", "fr"]);
 
     view! {
-        <p>Locale: <code class="font-bold">{locale}</code></p>
+        <p>Locale: <code class="font-bold">{move || locale.get().to_string()}</code></p>
     }
 }